* load full selinux db;
see "man 5 selabel_file"

SELABEL_OPT_SUBSET
	A non-null value for this option is interpreted as a path prefix, for
example "/etc". Only file context specifications with starting with a
first component that prefix matches the
	given prefix are loaded. This may increase lookup performance, however
any attempt to look up a path not starting with the given prefix may
fail. This optimization is no longer required due to the use of
file_contexts.bin files and is deprecated.